requirements.txt
Flask==3.1.2
numpy==2.3.3
pillow==11.3.0
PyYAML==6.0.2
tensorflow-cpu==2.20.0
